pyproject.toml
[tool.poetry]
name = "sysbot-helper-bot"
version = "0.1.0"
description = ""
authors = ["Magma5 <mosciskistan@outlook.com>"]
packages = [
    { include = "sysbot_helper", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Jinja2 = "^3.1.2"
Pillow = "^9.3.0"
PyYAML = "^6.0"
mss = "^7.0.1"
more-itertools = "^8.12.0"
py-cord = "2.3.2"
pydantic = "^1.10.4"
aiogram = "3.0.0b6"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.36"}
alembic = "^1.9.1"
asyncpg = "^0.27.0"
python-frontmatter = "^1.0.0"
iso8601 = "^1.1.0"
python-slugify = "^7.0.0"
markdownify = "^0.11.6"

[tool.poetry.scripts]
bot = "sysbot_helper:bot_main"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
black = "^22.12.0"
flake8 = "^6.0.0"
pytype = "^2022.12.15"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
